sampling fix

+++ dataset/EncodedDataset.py
@@ -114,7 +114,7 @@ class EncodedDataset(Dataset):
         for cluster_id in range(36):
             print("Dim cluster_labels:", self.cluster_labels.shape)
             cluster_samples = np.sum(self.cluster_labels == cluster_id)
-            samples_to_take = cluster_samples * weights[cluster_id]
+            samples_to_take = int(cluster_samples * weights[cluster_id])
 
             selected_indices = np.random.choice(
                 cluster_samples, size=samples_to_take, replace=False
